requirements.txt
# EduMi — Python dependencies
# Python 3.10+ | pip install -r requirements.txt

Django==4.2.9
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
django-compressor==4.4
django-extensions==3.2.3
django-browser-reload==1.21.0
Pillow>=11.0.0
cryptography>=42.0.0
openpyxl>=3.1.2
requests==2.31.0
python-dotenv>=1.0.0

